small refactoring, lint prettier
