+++ packages/grid/_modules_/grid/components/index.ts
@@ -1,5 +1,4 @@
 export * from './columnHeaders';
-export * from './filterPanel';
 export * from './icons';
 export * from './menu';
 export * from './styled-wrappers';
