+++ packages/grid/_modules_/grid/components/panel/filterPanel/FilterInputValue.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import TextField from '@material-ui/core/TextField';
-import { LoadIcon } from '../icons/index';
+import { LoadIcon } from '../../icons/index';
 import { FilterInputValueProps } from './FilterInputValueProps';
 
 const SUBMIT_FILTER_STROKE_TIME = 500;
