+++ packages/grid/_modules_/grid/components/panel/filterPanel/FilterInputValueProps.ts
@@ -1,4 +1,4 @@
-import { FilterItem } from '../../models/filterItem';
+import { FilterItem } from '../../../models/filterItem';
 
 export interface FilterInputValueProps {
   item: FilterItem;
