+++ packages/grid/_modules_/grid/components/panel/index.ts
@@ -1,3 +1,4 @@
 export * from './ColumnsPanel';
 export * from './Panel';
 export * from './PreferencesPanel';
+export * from './filterPanel';
