+++ packages/grid/_modules_/grid/models/filterOperator.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { FilterInputValueProps } from '../components/filterPanel/FilterInputValueProps';
+import { FilterInputValueProps } from '../components/panel/filterPanel/FilterInputValueProps';
 import { FilterItem } from './filterItem';
 import { CellParams } from './params/cellParams';
